builtinコマンドのNorm対応をしました。また、テストケースが正常に動作するように変更しました。
