+++ srcs/builtin_cmd/exit.c
@@ -6,28 +6,14 @@
 /*   By: yootaki <yootaki@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/04 11:41:46 by yootaki           #+#    #+#             */
-/*   Updated: 2021/10/01 22:30:38 by yootaki          ###   ########.fr       */
+/*   Updated: 2021/10/14 21:19:05 by yootaki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "builtin_cmd.h"
+#include "../../libft/libft.h"
 
 #define TOO_MANY_ARGUMENTS "minishell: exit: too many arguments\n"
-
-static int	is_str_digit(char *str)
-{
-	if (*str == '-' || *str == '+')
-		str++;
-	if (!*str)
-		return (0);
-	while (*str)
-	{
-		if (!ft_isdigit(*str))
-			return (0);
-		str++;
-	}
-	return (1);
-}
 
 static int	ft_ovcheck(int sign, long num, long next_num)
 {
@@ -85,15 +71,31 @@ int	calculate_exit_status(char *str)
 		return (num % 256);
 }
 
+int	determine_exit_status(t_cmd_lst *now, int count, char **args)
+{
+	if ((count == 0 && now->next->next->str[0] == ' ') \
+	|| longlong_over_check(args[0]))
+		return (255);
+	if (count == 1)
+		return (calculate_exit_status(args[0]));
+	if (count == 2 && is_str_digit(args[0]) \
+	&& !longlong_over_check(args[0]) && !is_str_digit(args[1]))
+		return (255);
+	else if (count == 2 && !is_str_digit(args[0]))
+		return (255);
+	return (1);
+}
+
 int	my_exit(t_cmd_lst *cmd)
 {
 	t_cmd_lst	*now;
 	char		*args[2];
 	int			count;
 
+	g_status = 0;
 	now = cmd->next->next;
 	if (now == cmd)
-		exit (0);
+		exit (g_status);
 	count = 0;
 	while (now != cmd)
 	{
@@ -103,20 +105,12 @@ int	my_exit(t_cmd_lst *cmd)
 		{
 			g_status = 1;
 			ft_putstr_fd(TOO_MANY_ARGUMENTS, STDERR_FILENO);
-			return (1);
+			return (g_status);
 		}
 		else if (count > 0)
 			args[count - 1] = now->str;
 		now = now->next;
 	}
-	if ((count == 0 && now->next->next->str[0] == ' ') || longlong_over_check(args[0]))
-		exit (255);
-	if (count == 1)
-		exit (calculate_exit_status(args[0]));
-	if (count == 2 && is_str_digit(args[0]) \
-	&& !longlong_over_check(args[0]) && !is_str_digit(args[1]))
-		exit (255);
-	else if (count == 2 && !is_str_digit(args[0]))
-		exit (255);
-	exit (1);
+	g_status = determine_exit_status(now, count, args);
+	exit(g_status);
 }
