+++ srcs/builtin_cmd/ft_atol.c
@@ -6,11 +6,11 @@
 /*   By: yootaki <yootaki@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/08 09:40:44 by yootaki           #+#    #+#             */
-/*   Updated: 2021/09/27 21:14:39 by yootaki          ###   ########.fr       */
+/*   Updated: 2021/10/14 22:21:38 by yootaki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../../libft/libft.h"
 #include "builtin_cmd.h"
 
 static int	ft_ovcheck(int sign, long num, long next_num)
