+++ srcs/expansion/redirect.c
@@ -6,7 +6,7 @@
 /*   By: yootaki <yootaki@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/06 23:08:43 by yootaki           #+#    #+#             */
-/*   Updated: 2021/10/02 22:48:12 by yootaki          ###   ########.fr       */
+/*   Updated: 2021/10/14 11:40:06 by yootaki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
