+++ testcase/scripts/test_builtins.sh
@@ -58,56 +58,56 @@ test_echo ()
 }
 
 # test_cd () {
-    run_test 'cd /.//'
-    run_test 'cd /./////././////'
-    run_test 'cd //'
-    run_test 'cd //...........//////'
-    run_test 'cd //'
-    run_test 'cd //'
-    run_test 'cd ///'
-    run_test 'cd //.'
-    run_test 'cd //./'
-    run_test 'cd //.///./././////'
-    run_test 'cd ///'
-    run_test 'cd ///.///'
-    run_test 'cd ////'
-    run_test 'cd ////./././//////'
-    run_test 'cd /////'
-    run_test 'cd /bin'
-    run_test 'cd /usr/bin'
-    run_test 'cd /home'
-    run_test 'cd //home'
-    run_test 'cd //./home'
-    run_test 'cd "//home"'
-    run_test 'cd $OLDPWD'
-    run_test 'cd $WWW'
-    run_test 'cd $WWW$WWW../../'
-    run_test 'cd $WWW./$WWW../$WWW'
-    run_test 'cd ../../../../../../../'
-    run_test 'cd ../.././.././.././.././.././../'
-    run_test '""c""d"" .'
-    run_test 'cd .'
-    run_test 'cd ..'
-    run_test 'cd ..""""'
-    run_test 'cd ""."".""""'
-    run_test 'cd ...'
-    run_test 'cd ../..'
-    run_test 'cd ../.'
-    run_test 'cd ./.'
-    run_test 'cd /'
-    run_test 'cd //'
-    run_test 'cd //home'
-    run_test 'cd "//home"'
-    run_test 'cd aaaaaa'
-    run_test 'cd ""'
-    run_test 'cd " "'
-    run_test 'cd ./'
-    run_test 'cd /.'
-    run_test 'cd /..'
-    run_test 'cd'
-    run_test 'cd                    '
-    run_test 'cd        ""    ""        '
-    run_test 'cd        " "        '
+    # run_test 'cd /.//'
+    # run_test 'cd /./////././////'
+    # run_test 'cd //'
+    # run_test 'cd //...........//////'
+    # run_test 'cd //'
+    # run_test 'cd //'
+    # run_test 'cd ///'
+    # run_test 'cd //.'
+    # run_test 'cd //./'
+    # run_test 'cd //.///./././////'
+    # run_test 'cd ///'
+    # run_test 'cd ///.///'
+    # run_test 'cd ////'
+    # run_test 'cd ////./././//////'
+    # run_test 'cd /////'
+    # run_test 'cd /bin'
+    # run_test 'cd /usr/bin'
+    # run_test 'cd /home'
+    # run_test 'cd //home'
+    # run_test 'cd //./home'
+    # run_test 'cd "//home"'
+    # run_test 'cd $OLDPWD'
+    # run_test 'cd $WWW'
+    # run_test 'cd $WWW$WWW../../'
+    # run_test 'cd $WWW./$WWW../$WWW'
+    # run_test 'cd ../../../../../../../'
+    # run_test 'cd ../.././.././.././.././.././../'
+    # run_test '""c""d"" .'
+    # run_test 'cd .'
+    # run_test 'cd ..'
+    # run_test 'cd ..""""'
+    # run_test 'cd ""."".""""'
+    # run_test 'cd ...'
+    # run_test 'cd ../..'
+    # run_test 'cd ../.'
+    # run_test 'cd ./.'
+    # run_test 'cd /'
+    # run_test 'cd //'
+    # run_test 'cd //home'
+    # run_test 'cd "//home"'
+    # run_test 'cd aaaaaa'
+    # run_test 'cd ""'
+    # run_test 'cd " "'
+    # run_test 'cd ./'
+    # run_test 'cd /.'
+    # run_test 'cd /..'
+    # run_test 'cd'
+    # run_test 'cd                    '
+    # run_test 'cd        ""    ""        '
+    # run_test 'cd        " "        '
     # run_test 'cd ~ ; pwd'
     # run_test 'cd ~/ ; pwd'
     # run_test 'cd ~/. ; pwd'
